fix(avaliacao): update param nota int to float

+++ 001/Estrurando_Aplicacoes_Web/aplicando_orientacao_a_objetos/oop_sabor_express/modelos/avaliacao.py
@@ -1,4 +1,4 @@
 class Avaliacao:
-    def __init__(self, cliente: str, nota: int):
+    def __init__(self, cliente: str, nota: float):
         self._cliente = cliente
         self._nota = nota
